requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
Pillow>=10.0.0
numpy>=1.26.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
